Version 1
This program calculates the percentage of shared kmer sequences between the Capsid and Rep genes of each genome of a file. It writes a file that contains a ranking of the genomes based on the similarity of kmer sequences between its Capsid and Rep genes with higher similarity earlier in the rank.